nlabosesamebot/requirements.txt
pysesameos2
discord.py